requirements.txt
streamlit
pillow
numpy
matplotlib
scikit-learn
scikit-image
